requirements.txt
streamlit==0.67.1
cufflinks==0.17.3
decorator==4.4.2
flasgger==0.9.5
Flask==1.1.2
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
MarkupSafe==1.1.1
matplotlib==3.3.2
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandas==1.1.2
pickleshare==0.7.5
Pillow==7.2.0
pip-check-reqs==2.1.1
plotly==4.10.0
plotly-express==0.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
statsmodels==0.12.0
urllib3==1.25.10
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2

